Build 116 Changes
Fabric Icons retrieval improvements (#84)
    - Moved `deneb-config` to `/config`
    - Removed the .woff file and swapped to inline CSS for font declarations (as this means we just need a single file to manage content)
    - Icons are now driven from `/config/fabric-icons.json`
    - `getIcons()` added to `core/config`
    - `bin/update-fabric-icons` script added to package, which will:
        - Parse the icons object in `/config/fabric-icons.json`
        - Call the Office UI Fabric Icons tool and download the generated .zip
        - Extract the generated inline CSS file and apply to `style/fabric-icons-inline-css`
    - Moved `initializeIcons` from API to `core/ui/fluent`

Modal dialog enhancements (#85)
    - Moved styles into `core/ui/modal`
    - Added height and width percentages to config rather than hard-coding, and set these to 95%
