+++ src/components/create/NewVisualDialogBody.tsx
@@ -10,7 +10,7 @@ import {
     modalDialogStackItemStyles,
     modalDialogStackItemWrapperStyles,
     modalDialogInnerStackTokens
-} from '../../config/styles';
+} from '../../core/ui/modal';
 import { state } from '../../store';
 import NewVisualDialogPivot from './NewVisualDialogPivot';
 import TemplateManagementPane from './TemplateManagementPane';
