+++ src/config/styles.ts
@@ -1,6 +1,3 @@
-import powerbi from 'powerbi-visuals-api';
-import IViewport = powerbi.IViewport;
-
 import {
     getFocusStyle,
     getTheme,
@@ -13,7 +10,6 @@ import {
     ITheme,
     mergeStyleSets,
     FontSizes,
-    FontWeights,
     ITextFieldStyles,
     IDetailsListStyles,
     IStyleSet,
@@ -33,12 +29,6 @@ export {
     exportPivotAssistiveTextStyles,
     exportPivotAssistiveToastTextStyles,
     exportPivotStyles,
-    modalDialogCloseIconStyles,
-    modalDialogContentStyles,
-    modalDialogStackStyles,
-    modalDialogStackItemStyles,
-    modalDialogStackItemWrapperStyles,
-    modalDialogInnerStackTokens,
     templateExportInfoStackTokens,
     templatePickerDropdownStyles,
     templatePickerItemListStyles,
@@ -197,77 +187,6 @@ const templatePickerItemListStyles = mergeStyleSets({
     },
     templatePickerDropdownStyles: Partial<IDropdownStyles> = {
         dropdown: { width: 300 }
-    };
-
-// Modal dialog layout
-const modalDialogContentStyles = (viewport: IViewport) => {
-        const dialogHeight = viewport.height * 0.8,
-            dialogWidth = viewport.width * 0.8;
-        return mergeStyleSets({
-            container: {
-                display: 'flex',
-                flexFlow: 'column nowrap',
-                overflowY: 'hidden',
-                alignItems: 'stretch',
-                height: dialogHeight,
-                width: dialogWidth
-            },
-            header: [
-                {
-                    flex: '1 1 auto',
-                    borderTop: `4px solid ${theme.palette.themePrimary}`,
-                    color: theme.palette.neutralPrimary,
-                    display: 'flex',
-                    alignItems: 'center',
-                    fontSize: FontSizes.xLargePlus,
-                    fontWeight: FontWeights.semibold,
-                    padding: '12px 12px 14px 24px'
-                }
-            ],
-            body: {
-                flex: '4 4 auto',
-                padding: '0 24px 0px 24px',
-                overflowY: 'hidden',
-                selectors: {
-                    p: { margin: '14px 0' },
-                    'p:first-child': { marginTop: 0 },
-                    'p:last-child': { marginBottom: 0 }
-                },
-                height: `calc(100% - 4px - 12px - 14px - ${FontSizes.xLargePlus} - 14px)`
-            }
-        });
-    },
-    modalDialogCloseIconStyles = {
-        root: {
-            color: theme.palette.neutralPrimary,
-            marginLeft: 'auto',
-            marginTop: '4px',
-            marginRight: '2px'
-        },
-        rootHovered: {
-            color: theme.palette.neutralDark
-        }
-    },
-    modalDialogStackStyles: IStackStyles = {
-        root: {
-            height: '100%'
-        }
-    },
-    modalDialogStackItemStyles: IStackItemStyles = {
-        root: {
-            display: 'flex'
-        }
-    },
-    modalDialogStackItemWrapperStyles: IStackItemStyles = {
-        root: {
-            display: 'flex',
-            minHeight: 0,
-            paddingTop: '10px'
-        }
-    },
-    modalDialogInnerStackTokens: IStackTokens = {
-        childrenGap: 15,
-        padding: 10
     };
 
 // Export dialog - pivot-level assistive text
