+++ src/core/ui/index.ts
@@ -3,7 +3,8 @@ export * as aria from './aria';
 export * as commandBar from './commandBar';
 export * as commands from './commands';
 export * as dom from './dom';
-export * as fill from './fill';
+export * as fluent from './fluent';
 export * as i18n from './i18n';
 export * as icons from './icons';
 export * as labels from './labels';
+export * as modal from './modal';
