--- style/fabric-icons.css
@@ -1,108 +1,0 @@
-/*
-  Your use of the content in the files referenced here is subject to the terms of the license at https://aka.ms/fabric-assets-license
-*/
-@font-face {
-    font-family: 'FabricMDL2Icons';
-    src: url('../fonts/fabric-icons.woff') format('woff');
-}
-
-.ms-Icon {
-    -moz-osx-font-smoothing: grayscale;
-    -webkit-font-smoothing: antialiased;
-    display: inline-block;
-    font-family: 'FabricMDL2Icons';
-    font-style: normal;
-    font-weight: normal;
-    speak: none;
-}
-
-.ms-Icon--BarChartVertical:before {
-    content: '\E9EC';
-}
-.ms-Icon--CalculatorAddition:before {
-    content: '\E948';
-}
-.ms-Icon--CalculatorSubtract:before {
-    content: '\E949';
-}
-.ms-Icon--Calendar:before {
-    content: '\E787';
-}
-.ms-Icon--Cancel:before {
-    content: '\E711';
-}
-.ms-Icon--CheckMark:before {
-    content: '\E73E';
-}
-.ms-Icon--ChevronDown:before {
-    content: '\E70D';
-}
-.ms-Icon--ChevronLeft:before {
-    content: '\E76B';
-}
-.ms-Icon--ChevronRight:before {
-    content: '\E76C';
-}
-.ms-Icon--CircleStopSolid:before {
-    content: '\F2DB';
-}
-.ms-Icon--Copy:before {
-    content: '\E8C8';
-}
-.ms-Icon--EditStyle:before {
-    content: '\EF60';
-}
-.ms-Icon--HalfAlpha:before {
-    content: '\E97E';
-}
-.ms-Icon--Help:before {
-    content: '\E897';
-}
-.ms-Icon--Info:before {
-    content: '\E946';
-}
-.ms-Icon--NumberSymbol:before {
-    content: '\F7AC';
-}
-.ms-Icon--OpenFile:before {
-    content: '\E8E5';
-}
-.ms-Icon--Page:before {
-    content: '\E7C3';
-}
-.ms-Icon--Play:before {
-    content: '\E768';
-}
-.ms-Icon--PlaybackRate1x:before {
-    content: '\EC57';
-}
-.ms-Icon--Refresh:before {
-    content: '\E72C';
-}
-.ms-Icon--Repair:before {
-    content: '\E90F';
-}
-.ms-Icon--Settings:before {
-    content: '\E713';
-}
-.ms-Icon--Share:before {
-    content: '\E72D';
-}
-.ms-Icon--ToggleRight:before {
-    content: '\F19F';
-}
-.ms-Icon--Unknown:before {
-    content: '\E9CE';
-}
-.ms-Icon--Zoom:before {
-    content: '\E71E';
-}
-.ms-Icon--ZoomIn:before {
-    content: '\E8A3';
-}
-.ms-Icon--ZoomOut:before {
-    content: '\E71F';
-}
-.ms-Icon--ZoomToFit:before {
-    content: '\F649';
-}
